reversed renaming themes actions and reducers
